Fix PHP 8.1 warnings

+++ src/Convert.php
@@ -48,9 +48,9 @@ class Convert
     public static function cmykValueToRgb(float $cyan, float $magenta, float $yellow, float $key): array
     {
         return [
-            255 * (1 - $cyan) * (1 - $key),
-            255 * (1 - $magenta) * (1 - $key),
-            255 * (1 - $yellow) * (1 - $key),
+            (int) (255 * (1 - $cyan) * (1 - $key)),
+            (int) (255 * (1 - $magenta) * (1 - $key)),
+            (int) (255 * (1 - $yellow) * (1 - $key)),
         ];
     }
